# backend/src/main/java/com/artwell/api/dto/UploadResult.java
package com.artwell.api.dto;

import java.util.UUID;

public record UploadResult(
        UUID documentId,
        UUID versionId,
        ValidationResult validation,
        DocumentDetail document
) {}