# backend/src/main/java/com/artwell/api/enums/DocumentTypeCode.java
package com.artwell.api.enums;

/** Коды по контракту OpenAPI (минимальный перечень Минстроя + OTHER). */
public enum DocumentTypeCode {
    AXIS_LAYOUT_ACT,
    ENGINEERING_NETWORKS_ACT,
    HIDDEN_WORKS_ACT,
    STRUCTURES_ACT,
    GEODESY_BASE_ACT,
    EQUIPMENT_DEFECTS_ACT,
    HYDROPNEUMATIC_TEST_ACT,
    FIRE_WATER_TEST_ACT,
    INCOMING_CONTROL_ACT,
    CONCRETE_JOURNAL,
    AUTHOR_SUPERVISION_JOURNAL,
    GENERAL_WORK_LOG,
    MATERIALS_INCOMING_JOURNAL,
    INSPECTION_PROTOCOL,
    SUPERVISION_ACT_NO_CONTACT,
    EXTENSION_DECISION,
    DESKTOP_INSPECTION_ACT,
    FIELD_INSPECTION_ACT,
    VIOLATION_ORDER,
    VIOLATION_REMEDIATION_NOTICE,
    SUPERVISION_DECISION,
    OTHER
}